More than 30 % speedup on Mjolnir for both B-field NTTs and X-field NTTs. Should be benchmark in Triton VM setting (proving, multithreaded NTTs) before merge though.

Speedup verified on both Mjolnir and Megingjord.