@jan-ferdinand 2025-06-06
`TVM_LDE_TRACE=cache cargo run --release -- --profile prove --program ~/tmp_spin.tasm --input 23`

## Baseline

<details>
<summary>Details</summary>
<pre>
### Triton VM – Prove                         419.29s    #Reps   Share  Category          399.2 GiB
├─trace execution                               3.61s        1   0.86%  (gen  –  4.62%)    +1.1 GiB
├─Fiat-Shamir: claim                            5.45µs       1   0.00%  (hash –  0.00%)      ±0 B  
├─derive additional parameters                443.08µs       1   0.00%                       ±0 B  
├─main tables                                 202.85s        1  48.38%                   +235.8 GiB
│ ├─create                                     40.65s        1   9.70%  (gen  – 52.00%)   +24.8 GiB
│ ├─pad                                         1.70s        1   0.41%  (gen  –  2.18%)    +4.5 MiB
│ │ ├─pad original tables                       1.10s        1   0.26%                     +3.9 MiB
│ │ └─fill degree-lowering table              597.93ms       1   0.14%                   +660.0 KiB
│ ├─LDE                                       101.05s        1  24.10%  (LDE  – 54.10%)  +213.4 GiB
│ │ ├─polynomial zero-initialization            1.87µs       1   0.00%                       ±0 B  
│ │ ├─interpolation                            15.50s        1   3.70%                    +23.8 GiB
│ │ ├─resize                                    5.01s        1   1.19%                   +189.4 GiB
│ │ ├─evaluation                               80.54s        1  19.21%                   +201.7 MiB
│ │ └─memoize                                 581.00ns       1   0.00%                       ±0 B  
│ ├─Merkle tree                                26.13s        1   6.23%                     +4.9 GiB
│ │ ├─leafs                                    23.63s        1   5.64%                     +2.3 GiB
│ │ │ └─hash rows                              23.63s        1   5.64%  (hash – 46.02%)    +2.3 GiB
│ │ └─Merkle tree                               2.36s        1   0.56%  (hash –  4.60%)    +5.2 GiB
│ ├─Fiat-Shamir                                25.45µs       1   0.00%  (hash –  0.00%)      ±0 B  
│ └─extend                                     32.21s        1   7.68%  (gen  – 41.21%)   +16.5 GiB
│   ├─initialize master table                 582.33ms       1   0.14%                    +16.5 GiB
│   ├─slice master table                        5.12µs       1   0.00%                       ±0 B  
│   ├─all tables                               31.33s        1   7.47%                    +44.8 MiB
│   └─fill degree lowering table              303.53ms       1   0.07%                   +192.0 KiB
├─aux tables                                   83.73s        1  19.97%                   +136.1 GiB
│ ├─LDE                                        61.35s        1  14.63%  (LDE  – 32.85%)  +148.6 GiB
│ │ ├─polynomial zero-initialization            1.34µs       1   0.00%                       ±0 B  
│ │ ├─interpolation                            10.32s        1   2.46%                    +16.6 GiB
│ │ ├─resize                                    3.49s        1   0.83%                   +131.9 GiB
│ │ ├─evaluation                               47.53s        1  11.34%                   +125.9 MiB
│ │ └─memoize                                 842.00ns       1   0.00%                       ±0 B  
│ ├─Merkle tree                                21.61s        1   5.15%                     +4.9 GiB
│ │ ├─leafs                                    19.02s        1   4.54%                     +2.4 GiB
│ │ │ └─hash rows                              19.02s        1   4.54%  (hash – 37.03%)    +2.4 GiB
│ │ └─Merkle tree                               2.46s        1   0.59%  (hash –  4.79%)    +4.1 GiB
│ └─Fiat-Shamir                               112.24µs       1   0.00%  (hash –  0.00%)      ±0 B  
├─quotient calculation (cached)                23.72s        1   5.66%  (CC   – 62.06%)    +1.5 GiB
│ ├─zerofier inverse                            6.50s        1   1.55%                     +2.7 GiB
│ └─evaluate AIR, compute quotient codeword    17.13s        1   4.09%                     +1.5 GiB
├─quotient LDE                                 24.37s        1   5.81%  (LDE  – 13.05%)    +8.4 GiB
├─hash rows of quotient segments                1.54s        1   0.37%  (hash –  3.00%)    +2.5 GiB
├─Merkle tree                                   2.34s        1   0.56%  (hash –  4.56%)    +5.6 GiB
├─out-of-domain rows                           22.19s        1   5.29%                   +137.3 MiB
├─Fiat-Shamir                                  69.24µs       1   0.00%  (hash –  0.00%)      ±0 B  
├─linear combination                           22.07s        1   5.26%                     +2.5 GiB
│ ├─main                                        1.83s        1   0.44%  (CC   –  4.80%)   +76.1 MiB
│ ├─aux                                         1.63s        1   0.39%  (CC   –  4.26%)  +169.1 MiB
│ └─quotient                                    9.72s        1   2.32%  (CC   – 25.42%)    +0.1 GiB
├─DEEP                                          4.16s        1   0.99%                     +4.5 GiB
│ ├─main&aux curr row                           1.37s        1   0.33%                     +1.5 GiB
│ ├─main&aux next row                           1.39s        1   0.33%                     +1.6 GiB
│ └─segmented quotient                          1.40s        1   0.33%                     +1.5 GiB
├─combined DEEP polynomial                      1.33s        1   0.32%                     -2.1 GiB
│ └─sum                                         1.32s        1   0.32%  (CC   –  3.47%)    -2.1 GiB
├─FRI                                           9.28s        1   2.21%                    +65.6 MiB
└─open trace leafs                            839.57µs       1   0.00%                       ±0 B  
</pre>
</details>

```
### Categories
LDE    186.77s  44.54%
gen     78.18s  18.65%
hash    51.36s  12.25%
CC      38.23s   9.12%

Clock frequency is 6716 Hz (2816027 clock cycles / (419287 ms / 1 iterations))
Optimal clock frequency is 20006 Hz (8388608 padded height / (419287 ms / 1 iterations))
FRI domain length is 2^26
```

## With 585a051 (No Allocation)

<details>
<summary>Details</summary>
<pre>
### Triton VM – Prove                         406.74s    #Reps   Share  Category          399.6 GiB
├─trace execution                               3.63s        1   0.89%  (gen  –  4.72%)    +1.1 GiB
├─Fiat-Shamir: claim                            6.18µs       1   0.00%  (hash –  0.00%)      ±0 B  
├─derive additional parameters                440.33µs       1   0.00%                       ±0 B  
├─main tables                                 200.01s        1  49.17%                   +235.9 GiB
│ ├─create                                     38.98s        1   9.58%  (gen  – 50.74%)   +24.8 GiB
│ ├─pad                                         1.71s        1   0.42%  (gen  –  2.22%)    +2.8 MiB
│ │ ├─pad original tables                       1.10s        1   0.27%                     -1.4 MiB
│ │ └─fill degree-lowering table              606.80ms       1   0.15%                     +4.2 MiB
│ ├─LDE                                        99.66s        1  24.50%  (LDE  – 55.00%)  +213.4 GiB
│ │ ├─polynomial zero-initialization            1.86µs       1   0.00%                       ±0 B  
│ │ ├─interpolation                            15.88s        1   3.90%                    +23.9 GiB
│ │ ├─resize                                    5.02s        1   1.23%                   +189.4 GiB
│ │ ├─evaluation                               78.76s        1  19.36%                   +234.9 MiB
│ │ └─memoize                                 732.00ns       1   0.00%                       ±0 B  
│ ├─Merkle tree                                26.04s        1   6.40%                     +4.8 GiB
│ │ ├─leafs                                    23.52s        1   5.78%                     +2.3 GiB
│ │ │ └─hash rows                              23.52s        1   5.78%  (hash – 46.56%)    +2.3 GiB
│ │ └─Merkle tree                               2.39s        1   0.59%  (hash –  4.73%)    +5.5 GiB
│ ├─Fiat-Shamir                                26.55µs       1   0.00%  (hash –  0.00%)      ±0 B  
│ └─extend                                     32.51s        1   7.99%  (gen  – 42.32%)   +16.5 GiB
│   ├─initialize master table                 577.30ms       1   0.14%                    +16.5 GiB
│   ├─slice master table                        5.11µs       1   0.00%                       ±0 B  
│   ├─all tables                               31.63s        1   7.78%                    +38.2 MiB
│   └─fill degree lowering table              304.82ms       1   0.07%                     +8.0 KiB
├─aux tables                                   82.69s        1  20.33%                   +137.4 GiB
│ ├─LDE                                        60.68s        1  14.92%  (LDE  – 33.49%)  +148.6 GiB
│ │ ├─polynomial zero-initialization            1.62µs       1   0.00%                       ±0 B  
│ │ ├─interpolation                            10.26s        1   2.52%                    +16.6 GiB
│ │ ├─resize                                    3.50s        1   0.86%                   +131.9 GiB
│ │ ├─evaluation                               46.92s        1  11.54%                   +108.5 MiB
│ │ └─memoize                                 791.00ns       1   0.00%                       ±0 B  
│ ├─Merkle tree                                21.24s        1   5.22%                     +4.9 GiB
│ │ ├─leafs                                    18.73s        1   4.60%                     +2.4 GiB
│ │ │ └─hash rows                              18.73s        1   4.60%  (hash – 37.07%)    +2.4 GiB
│ │ └─Merkle tree                               2.38s        1   0.59%  (hash –  4.72%)    +4.1 GiB
│ └─Fiat-Shamir                               134.18µs       1   0.00%  (hash –  0.00%)      ±0 B  
├─quotient calculation (cached)                23.37s        1   5.75%  (CC   – 65.65%)    +1.5 GiB
│ ├─zerofier inverse                            6.49s        1   1.60%                     +1.1 GiB
│ └─evaluate AIR, compute quotient codeword    16.77s        1   4.12%                     +1.5 GiB
├─quotient LDE                                 20.84s        1   5.12%  (LDE  – 11.50%)    +8.2 GiB
├─hash rows of quotient segments                1.11s        1   0.27%  (hash –  2.20%)    +2.5 GiB
├─Merkle tree                                   2.39s        1   0.59%  (hash –  4.72%)    +5.5 GiB
├─out-of-domain rows                           21.92s        1   5.39%                   +152.3 MiB
├─Fiat-Shamir                                  67.93µs       1   0.00%  (hash –  0.00%)      ±0 B  
├─linear combination                           17.73s        1   4.36%                     +2.5 GiB
│ ├─main                                        1.63s        1   0.40%  (CC   –  4.57%)   +75.9 MiB
│ ├─aux                                         1.43s        1   0.35%  (CC   –  4.02%)  +170.1 MiB
│ └─quotient                                    7.74s        1   1.90%  (CC   – 21.74%)    +0.1 GiB
├─DEEP                                          4.17s        1   1.03%                     +4.5 GiB
│ ├─main&aux curr row                           1.37s        1   0.34%                     +1.5 GiB
│ ├─main&aux next row                           1.40s        1   0.34%                     +1.5 GiB
│ └─segmented quotient                          1.41s        1   0.35%                     +1.5 GiB
├─combined DEEP polynomial                      1.43s        1   0.35%                     -2.1 GiB
│ └─sum                                         1.43s        1   0.35%  (CC   –  4.02%)    -2.1 GiB
├─FRI                                           9.26s        1   2.28%                    +45.6 MiB
└─open trace leafs                            842.89µs       1   0.00%                       ±0 B  
</pre>
</details>

```
### Categories
LDE    181.19s  44.55%
gen     76.82s  18.89%
hash    50.52s  12.42%
CC      35.60s   8.75%

Clock frequency is 6923 Hz (2816027 clock cycles / (406744 ms / 1 iterations))
Optimal clock frequency is 20623 Hz (8388608 padded height / (406744 ms / 1 iterations))
FRI domain length is 2^26
```

## With 572a3aa (Precompute Twiddle Factors)

<details>
<summary>Details</summary>
<pre>
### Triton VM – Prove                         409.94s    #Reps   Share  Category          399.9 GiB
├─trace execution                               3.61s        1   0.88%  (gen  –  4.83%)    +1.1 GiB
├─Fiat-Shamir: claim                            6.04µs       1   0.00%  (hash –  0.00%)      ±0 B  
├─derive additional parameters                446.39µs       1   0.00%                       ±0 B  
├─main tables                                 202.05s        1  49.29%                   +238.6 GiB
│ ├─create                                     37.50s        1   9.15%  (gen  – 50.14%)   +24.7 GiB
│ ├─pad                                         1.68s        1   0.41%  (gen  –  2.25%)  -820.0 KiB
│ │ ├─pad original tables                       1.10s        1   0.27%                     -1.8 MiB
│ │ └─fill degree-lowering table              584.82ms       1   0.14%                   +972.0 KiB
│ ├─LDE                                       103.90s        1  25.35%  (LDE  – 55.53%)  +215.8 GiB
│ │ ├─polynomial zero-initialization            1.64µs       1   0.00%                       ±0 B  
│ │ ├─interpolation                            14.99s        1   3.66%                    +26.1 GiB
│ │ ├─resize                                    5.01s        1   1.22%                   +189.4 GiB
│ │ ├─evaluation                               83.90s        1  20.47%                   +192.2 MiB
│ │ └─memoize                                 752.00ns       1   0.00%                       ±0 B  
│ ├─Merkle tree                                25.86s        1   6.31%                     +4.9 GiB
│ │ ├─leafs                                    23.32s        1   5.69%                     +2.3 GiB
│ │ │ └─hash rows                              23.32s        1   5.69%  (hash – 46.50%)    +2.3 GiB
│ │ └─Merkle tree                               2.39s        1   0.58%  (hash –  4.77%)    +5.2 GiB
│ ├─Fiat-Shamir                                25.36µs       1   0.00%  (hash –  0.00%)      ±0 B  
│ └─extend                                     31.99s        1   7.80%  (gen  – 42.77%)   +16.5 GiB
│   ├─initialize master table                 609.20ms       1   0.15%                    +16.5 GiB
│   ├─slice master table                        4.10µs       1   0.00%                       ±0 B  
│   ├─all tables                               31.07s        1   7.58%                    +26.1 MiB
│   └─fill degree lowering table              302.99ms       1   0.07%                   +104.0 KiB
├─aux tables                                   84.38s        1  20.58%                   +136.1 GiB
│ ├─LDE                                        62.51s        1  15.25%  (LDE  – 33.41%)  +148.7 GiB
│ │ ├─polynomial zero-initialization            1.38µs       1   0.00%                       ±0 B  
│ │ ├─interpolation                            10.43s        1   2.54%                    +16.6 GiB
│ │ ├─resize                                    3.50s        1   0.85%                   +131.9 GiB
│ │ ├─evaluation                               48.58s        1  11.85%                   +176.2 MiB
│ │ └─memoize                                 430.00ns       1   0.00%                       ±0 B  
│ ├─Merkle tree                                21.10s        1   5.15%                     +4.9 GiB
│ │ ├─leafs                                    18.60s        1   4.54%                     +2.3 GiB
│ │ │ └─hash rows                              18.60s        1   4.54%  (hash – 37.08%)    +2.3 GiB
│ │ └─Merkle tree                               2.37s        1   0.58%  (hash –  4.72%)    +5.1 GiB
│ └─Fiat-Shamir                               113.22µs       1   0.00%  (hash –  0.00%)      ±0 B  
├─quotient calculation (cached)                23.37s        1   5.70%  (CC   – 66.22%)    +1.5 GiB
│ ├─zerofier inverse                            6.48s        1   1.58%                     +1.1 GiB
│ └─evaluate AIR, compute quotient codeword    16.77s        1   4.09%                     +1.5 GiB
├─quotient LDE                                 20.69s        1   5.05%  (LDE  – 11.06%)    +6.9 GiB
├─hash rows of quotient segments                1.11s        1   0.27%  (hash –  2.22%)    +2.5 GiB
├─Merkle tree                                   2.36s        1   0.58%  (hash –  4.71%)    +4.1 GiB
├─out-of-domain rows                           22.08s        1   5.39%                   +116.1 MiB
├─Fiat-Shamir                                  67.67µs       1   0.00%  (hash –  0.00%)      ±0 B  
├─linear combination                           17.30s        1   4.22%                     +2.5 GiB
│ ├─main                                        1.64s        1   0.40%  (CC   –  4.65%)   +91.5 MiB
│ ├─aux                                         1.42s        1   0.35%  (CC   –  4.02%)  +170.4 MiB
│ └─quotient                                    7.53s        1   1.84%  (CC   – 21.33%)    +0.1 GiB
├─DEEP                                          4.22s        1   1.03%                     +4.5 GiB
│ ├─main&aux curr row                           1.38s        1   0.34%                     +1.5 GiB
│ ├─main&aux next row                           1.44s        1   0.35%                     +1.5 GiB
│ └─segmented quotient                          1.41s        1   0.34%                     +1.5 GiB
├─combined DEEP polynomial                      1.33s        1   0.32%                     -2.1 GiB
│ └─sum                                         1.33s        1   0.32%  (CC   –  3.78%)    -2.1 GiB
├─FRI                                           9.24s        1   2.25%                    +52.9 MiB
└─open trace leafs                            782.26µs       1   0.00%                       ±0 B  
</pre>
</details>

```
### Categories
LDE    187.10s  45.64%
gen     74.79s  18.24%
hash    50.16s  12.24%
CC      35.28s   8.61%

Clock frequency is 6869 Hz (2816027 clock cycles / (409943 ms / 1 iterations))
Optimal clock frequency is 20462 Hz (8388608 padded height / (409943 ms / 1 iterations))
FRI domain length is 2^26
```


I'd expected a larger speedup overall, to be honest. It's probably also worth to run the proving step more than once, as it seems to be somewhat noisy. For example, the `hash` category shouldn't change between the versions.